Add `get_valid_nodes` function
This returns a list of nodes IDs which have status data in redis.
This isn't a method of the NodeControl class, which is instantiated for
an individual node.

It might be useful in a construct something like

```
import nodeControl as nc

node_interfaces = []

for node in nc.get_valid_nodes():
    node_interfaces += nc.NodeControl(node)

for node_interface in node_interfaces:
    print node_interface.get_power_status()
```

Maybe this construct should be part of the control package somewhere...